requirement.txt.txt
streamlit==1.32.0
pandas==2.1.4
mysql-connector-python==9.6.0
plotly==5.20.0
numpy==1.26.4